Add files via upload
Here is a simple linear regression model that predicts house prices based on the given data set and its features. If the dataset could not be accessed, try using this URL, which sends you directly to the dataset used and can be downloaded from there (URL: https://www.kaggle.com/datasets/shree1992/housedata).